requirements.txt
Django>=5.2,<5.3
djangorestframework
psycopg[binary]
python-decouple
django-filter
